Okay, literally just had to change one thing
Why the hell does unreal's basic damn config settings not update when I fix up the redirectors, jesus christ when Epic games made this it was made with sticks and string, it's like they never thought someone might reorganize the game instance or the gamemode, jesus christ. That doesn't even cover the fact that when I started moving the files, the async loading screen thing made the game crash every time the editors booted up, because OH NO the loading screen render wasn't where it was supposed to be so I had to manually change that. I just thought that was the plugins being stupid....NOPE it was just Epic games throwing something together and hoping certain things just don't happen. I rant and I rant because I hate this stupid damn program, but we made the project in this, and no matter what program I find I will likely find something like this that just pisses me off jesus christ on a fucking bicycle.

Anyway problem should be fixed.