Docks Changes and Infector Boss Fixes
Okay, so I moved the infector boss to a secluded island instead of under the map, because then you can see the part. You can't get the part, but you can fight and kill the boss, and it doesn't break.

I added back in the fridge electricity changing when the vacuum tubes are broken, but i didn't have it reset when the collectable is collected, only because it doesn't really make much sense for the electricity to turn back on when the tubes are still blown.

The infector boss now works. I changed out the origin spawning to a dedicated origin actor for it to go to, and fixed a bug that had it doubling up the transformation and sending the target destination into the wrong direction.

I have also made it so that the teleporter should be unlocked after only 3 collectables, but I have not yet tested this.